moved delete request to api folder
